Further menu factory decomposition
